requirements.txt
psutil==5.9.6
slack_sdk==3.23.0
python-dotenv==1.0.0
